requirements.txt
fastapi
uvicorn[standard]==0.23.1
psycopg[binary]
SQLAlchemy>=2.2,<3
alembic==1.11.1
python-dotenv==1.0.0
asyncpg